requirement.txt
appdirs==1.4.4
autopep8==1.5.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coloredlogs==14.0
cryptography==3.1
cssselect==1.1.0
decorator==4.4.2
humanfriendly==8.2
idna==2.10
lxml==4.5.2
lyricwikia==0.1.11
mutagen==1.45.1
pathlib==1.0.1
pycodestyle==2.6.0
pycparser==2.20
python-telegram-bot==12.8
pytube3==9.6.4
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
spotdl==2.2.2
spotipy==2.14.0
toml==0.10.1
tornado==6.0.4
tqdm==4.48.2
typing-extensions==3.7.4.3
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.25.10
youtube-dl==2020.7.28
